requirements.txt
# ==========================================
# Core ML & Deep Learning
# ==========================================
torch>=2.0.0          # Основной фреймворк
torchvision>=0.15.0   # Трансформации изображений и базовые утилиты
timm>=0.9.0           # Vision Transformers (ViT) и другие SOTA модели

# ==========================================
# Data Processing & Image Manipulation
# ==========================================
pandas>=2.0.0         # Чтение и обработка train.csv
Pillow>=9.0.0         # Базовая работа с изображениями (RGB конвертация, загрузка)
opencv-python>=4.8.0  # Продвинутая обработка изображений (cv2)

# ==========================================
# External Modules (Segmentation / YOLO)
# ==========================================
ultralytics>=8.0.0    # Для запуска YOLOv11-12 из внешнего репозитория

# ==========================================
# Optional / Recommended for Production
# ==========================================
faiss-cpu>=1.7.4    # Раскомментируй для быстрого поиска (Векторная БД) вместо чистого PyTorch
# faiss-gpu>=1.7.4    # Если поиск будет крутиться на GPU

# ═══════════════════════════════════════════════════════════════
# requirements.txt
# Сегментация брюха тритона - RTX 3060 Ti
# ═══════════════════════════════════════════════════════════════

# Обработка изображений
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
scikit-image>=0.21.0

# Аугментация данных
albumentations>=1.3.0

# Визуализация и прогресс
matplotlib>=3.7.0
tqdm>=4.65.0

# Интерфейс приложения
gradio>=3.40.0

# Разметка данных
labelme>=5.4.0

# Примечание: PyTorch устанавливается отдельно с CUDA
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118